Changed to bigdecimal. Floating point arithmetic was getting out of hand

+++ src/main/kotlin/market/LimitOrder.kt
@@ -1,10 +1,11 @@
 package market
 
+import java.math.BigDecimal
 import java.util.*
 
 data class LimitOrder(
     val side: String,
-    val quantity: Double,
+    val quantity: BigDecimal,
     val price: Double,
     val pair: String,
     val orderTimestamp: Long = System.currentTimeMillis(),
